CDI_Backend/requirements.txt
# CDI Backend Requirements
# Core FastAPI and web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0

# HTTP requests
requests==2.31.0

# AI/OpenAI client
openai==1.3.7

# Optional development dependencies
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0
